+++ lib/avx10_t1/mb_mgr_avx10_t1.c
@@ -398,6 +398,9 @@ submit_job_docsis256_sec_crc_dec_vaes_avx512(MB_MGR_DOCSIS_AES_OOO *state, IMB_J
 /* AES-NIA5 */
 #define GENERATE_HQP_AES generate_hqp_vaes_avx512
 
+/* SNOW5G-NIA4 */
+#define GENERATE_HQP_SNOW5G generate_hqp_snow5g_sse
+
 /* ====================================================================== */
 
 static void