+++ lib/avx512_t1/mb_mgr_avx512_t1.c
@@ -358,6 +358,9 @@ submit_aes_docsis256_dec_crc32_avx512(MB_MGR_DOCSIS_AES_OOO *state, IMB_JOB *job
 /* AES-NIA5 */
 #define GENERATE_HQP_AES generate_hqp_aes_sse
 
+/* SNOW5G-NIA4 */
+#define GENERATE_HQP_SNOW5G generate_hqp_snow5g_sse
+
 /* ====================================================================== */
 
 static void