+++ lib/include/arch_sse_type1.h
@@ -369,6 +369,9 @@ IMB_JOB *
 snow_5g_sse(IMB_JOB *job);
 
 IMB_JOB *
+snow_5g_nia4_sse(IMB_JOB *job);
+
+IMB_JOB *
 submit_job_snow3g_uea2_sse(MB_MGR_SNOW3G_OOO *state, IMB_JOB *job);
 IMB_JOB *
 flush_job_snow3g_uea2_sse(MB_MGR_SNOW3G_OOO *state);