+++ test/kat-app/main.c
@@ -127,6 +127,8 @@ extern int
 aes_nca5_test(struct IMB_MGR *mb_mgr);
 extern int
 zuc_nca6_test(struct IMB_MGR *mb_mgr);
+extern int
+snow5g_nia4_test(struct IMB_MGR *mb_mgr);
 
 typedef int (*imb_test_t)(struct IMB_MGR *mb_mgr);
 
@@ -193,6 +195,7 @@ struct imb_test tests[] = {
         { .str = "AES_NEA5", .fn = aes_nea5_test, .enabled = 1 },
         { .str = "AES_NCA5", .fn = aes_nca5_test, .enabled = 1 },
         { .str = "ZUC_NCA6", .fn = zuc_nca6_test, .enabled = 1 },
+        { .str = "SNOW5G_NIA4", .fn = snow5g_nia4_test, .enabled = 1 }
 };
 
 static int self_test_corrupt = 0;