+++ test/xvalid-app/xvalid_tests.cmake
@@ -153,7 +153,8 @@ set(HASH_ALGOS
     shake-128
     shake-256
     aes-nia5
-    zuc-nia6)
+    zuc-nia6
+    snow5g-nia4)
 
 # hash short tests
 foreach(ALGO ${HASH_ALGOS})