+++ ReleaseNotes.md
@@ -29,6 +29,7 @@ Library
 - Added AES-NCA5 SSE type 1 and AVX512 type 2 implementations.
 - Added ZUC-NIA6 SSE type 1 implementation.
 - Added ZUC-NCA6 SSE type 1 and SSE type 3 implementations.
+- Added SNOW5G-NIA4 SSE type 1 implementation.
 
 Test Applications
 - Added `--offset` parameter to `imb-xvalid` application.